split channel matcher logic.
Remove fom FDC3 service into a new service.
Simplify the Component UI.
