+++ src/services/FDC3ChannelMatcher/FDC3ChannelMatcher.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>FDC3ChannelMatcher</title>
+  </head>
+
+  <body>
+    <h1>FDC3ChannelMatcher</h1>
+  </body>
+  <script src="FDC3ChannelMatcher.js"></script>
+</html>
