+++ src/services/FDC3ChannelMatcher/config.json
@@ -1,0 +1,23 @@
+{
+  "comment": "Houses config for any custom services that you'd like to import into Finsemble.",
+  "services": {
+    "FDC3ChannelMatcher": {
+      "bootParams": {
+        "stage": "preuser",
+        "stopOnFailure": false,
+        "autoStart": true,
+        "customFailureMessage": null,
+        "timeout": 10000,
+        "dependencies": [
+          "FDC3"
+        ]
+      },
+      "useWindow": true,
+      "active": true,
+      "name": "FDC3ChannelMatcher",
+      "file": "$applicationRoot/services/FDC3ChannelMatcher/FDC3ChannelMatcherService.js",
+      "visible": false,
+      "html": "$applicationRoot/services/FDC3ChannelMatcher/FDC3ChannelMatcher.html"
+    }
+  }
+}
